Move NIST DB fixture to conftest.py
Required since it is now used within 2 test files

(STEPS-174)
