+++ tests/test_database.py
@@ -31,11 +31,6 @@ def get_test_data_dir():
     return Path(os.path.abspath(__file__)).parents[0] / "data"
 
 
-@pytest.fixture(scope="session")
-def get_databases_dir():
-    return Path(os.path.abspath(__file__)).parents[1] / "databases"
-
-
 LANGUAGES = {
     'bg_BG': 'Bulgarian',
     'cs_CZ': 'Czech',
@@ -71,12 +66,6 @@ def guard_locale():
     old_locale = locale.setlocale(locale.LC_NUMERIC)
     yield old_locale
     locale.setlocale(locale.LC_NUMERIC, old_locale)
-
-
-@pytest.fixture(scope="session")
-def nist_database(get_databases_dir):
-    database = Database(str(get_databases_dir / "nist" / "database_nist_260822.xml"))
-    return database
 
 
 def try_set_locale(loc):
